auth-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic[email]
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.1.2
pytest==7.4.4
pytest-asyncio==0.23.3
requests==2.31.0